Add a task to push/pull POT/PO files from the Zanata server
New task that may affects the workflow would be 'potupload'.
Once all string changes is frozen, this task would works to upload
the source strings file in the gettext format to the Zanata server,
the web-based translation platform.
And can ask translators to get it in various languages.

Once it has been done or time to make a new release,
'dist' task would pulls the latest translations from the Zanata
server to include them in the release.

Note that uploading into the Zanata server requires an account on it
and zanata.ini file with the API key to automate the process.