RHMAP-9554 Add check for whether token is required
Currently the token which is provided to the environment `create`
and `update` commands are completely optional. This should not be
the case as any environment configured to target `openshift3` type
MBaaS should have to provide a token. Although this will be enforced
on the server-side it is also nice to have the requirement client-side.

This adds a check on the `create` and `update` commands for environments
to check the type of the MBaaS target. If the MBaaS type is `openshift3`
then a `token` is required otherwise an error will be returned.