RHMAP-9554 Update tests for environment target check
Now that the environment `create` and `update` methods perform a check
on the target type before performing an action the tests in fhc need
to be updated to account for the new api calls being made.

This adds extra calls to the `nock` interceptor list for checking the
target type.